+++ sqlexecutor/sqlexecutor_darwin_test.go
@@ -1,0 +1,5 @@
+//go:build darwin
+
+package sqlexecutor
+
+var sqliteDatabaseFilename = "/tmp/sqlite/G2C.db"
