+++ sqlexecutor/sqlexecutor_linux_test.go
@@ -1,0 +1,5 @@
+//go:build linux
+
+package sqlexecutor
+
+var sqliteDatabaseFilename = "/tmp/sqlite/G2C.db"
