+++ sqlexecutor/sqlexecutor_windows_test.go
@@ -1,0 +1,5 @@
+//go:build windows
+
+package sqlexecutor
+
+var sqliteDatabaseFilename = `C:\Temp\sqlite\G2C.db`
